% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_functions.R
\name{bounds.r}
\alias{bounds.r}
\title{Obtain the range of sample correlations for a given population correlation and sample size}
\usage{
bounds.r(samples, level = 0.95)
}
\arguments{
\item{level}{Confidence level (0 to 1), default .95}

\item{pop_r}{Population correlation}

\item{n}{Sample size}
}
\description{
Obtain the range of sample correlations for a given population correlation and sample size
}
