% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_functions.R
\name{ci.r}
\alias{ci.r}
\title{Obtain the confidence interval for a given sample correlations}
\usage{
ci.r(r, n, level = 0.95, show.message = TRUE)
}
\arguments{
\item{r}{Sample correlation}

\item{n}{Sample size}

\item{level}{Confidence level (0 to 1), default .95}

\item{show.message}{Show console message describing CI}
}
\description{
Obtain the confidence interval for a given sample correlations
}
