% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions_r.R
\name{get.r.samples}
\alias{get.r.samples}
\title{Calculate a number of sample correlations based on a specified population correlation}
\usage{
get.r.samples(pop.r = NA, n, number.of.samples = 10)
}
\arguments{
\item{pop.r}{Population correlation. Do not use pop.data is you provide this value.}

\item{n}{Sample size for all samples If you use n, do not use n.min or n.max.}

\item{number.of.samples}{Number of samples to obtain}
}
\value{
Data frame with sample correlations
}
\description{
Calculate a number of sample correlations based on a specified population correlation
}
\examples{
get.r.samples(pop.r = .35,n=100)
}
