% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageDetails.R
\docType{package}
\name{intervalTraining}
\alias{intervalTraining}
\alias{intervalTraining-package}
\title{Learn about confidence intervals and prediction intervals}
\description{
Learn about confidence intervals and prediction intervals
\tabular{ll}{
Package: \tab intervalTraining\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2019-05-06\cr
License: \tab MIT\cr
}
}
\author{
\tabular{ll}{
Author: \tab David J. Stanley \email{dstanley@uoguelph.ca}\cr
Maintainer: \tab David J. Stanley \email{dstanley@uoguelph.ca}
}
}
