% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions_r.R
\name{percent.true}
\alias{percent.true}
\title{Calculate the percentage of rows that are true in a column}
\usage{
percent.true(x)
}
\arguments{
\item{x}{the column to be examined}
}
\value{
The percent equal to TRUE
}
\description{
Calculate the percentage of rows that are true in a column
}
