% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_functions.R
\name{pop_r_sampling_bounds}
\alias{pop_r_sampling_bounds}
\title{Obtain the range of sample correlations for a given population correlation and sample size}
\usage{
pop_r_sampling_bounds(pop_r, n, level = 0.95)
}
\arguments{
\item{pop_r}{Population correlation}

\item{n}{Sample size}

\item{level}{Confidence level (0 to 1), default .95}
}
\description{
Obtain the range of sample correlations for a given population correlation and sample size
}
