% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-histotext.R
\name{stat_histotext}
\alias{stat_histotext}
\title{This is the documentation for stat_histotext}
\usage{
stat_histotext(mapping = NULL, data = NULL, geom = "text",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{asdfasd}

\item{data}{adfasd}

\item{geom}{asdfasd}

\item{position}{asdfasd}

\item{na.rm}{adsfasdf}

\item{show.legend}{asdfads}

\item{inherit.aes}{adfasd}

\item{...}{adfasdf}
}
\description{
This is the documentation for stat_histotext
}
