% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-piplot.R
\name{stat_piplot}
\alias{stat_piplot}
\title{This is the documentation for stat_catseye}
\usage{
stat_piplot(mapping = NULL, data = NULL, geom = "polygon",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, pop.r, n, pi.LL, pi.UL, ...)
}
\arguments{
\item{mapping}{asdfasd}

\item{data}{adfasd}

\item{geom}{asdfasd}

\item{position}{asdfasd}

\item{na.rm}{adsfasdf}

\item{show.legend}{asdfads}

\item{inherit.aes}{adfasd}

\item{...}{adfasdf}
}
\description{
This is the documentation for stat_catseye
}
