% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplingFunctions_r.R
\name{get.r.samples}
\alias{get.r.samples}
\title{Calculate a number of sample correlations based on a specified population correlation}
\usage{
get.r.samples(pop_r = NA, n, number_samples = 10,
  number_decimals = 2)
}
\arguments{
\item{pop_r}{Population correlation. Do not use pop.data is you provide this value.}

\item{n}{Sample size for all samples If you use n, do not use n.min or n.max.}

\item{number_samples}{Number of samples to obtain}

\item{number_decimals}{Number of decimals to report in returned data frame}
}
\value{
Data frame with sample correlations
}
\description{
Calculate a number of sample correlations based on a specified population correlation
}
\examples{
get_r_samples(pop_r = .35, n = 100)
}
