% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-ciplot.R
\name{stat_ciplot}
\alias{stat_ciplot}
\title{This is the documentation for stat_ciplot (confidence interval plot)}
\usage{
stat_ciplot(mapping = NULL, data = NULL, position = "identity",
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE, pop.r, n,
  ci.LL, ci.UL, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping}

\item{data}{The data to be displayed in this layer}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().}

\item{pop.r}{Population correlation}

\item{n}{Sample size}

\item{ci.LL}{Confidence interval lower limit}

\item{ci.UL}{Confidence interval upper limit}

\item{...}{Other arguments}
}
\description{
This is the documentation for stat_ciplot (confidence interval plot)
}
